Added delete command; Added give to player and slot command functions; Filled in tab completes

+++ io/github/bananapuncher714/cartographer/core/internal/Util_1_8.java
@@ -106,6 +106,6 @@ public class Util_1_8 implements GeneralUtil {
 	
 	@Override
 	public CrossVersionMaterial getBlockType( Block block ) {
-		return new CrossVersionMaterial( block.getType(), ( int ) block.getData() );
+		return new CrossVersionMaterial( block.getType(), block.getData() );
 	}
 }
